Remove LocationFetcherProvider and change tests.
